[refactor] book 관련 dto 파일구조 변경
